requirements.txt
django==1.8.6
xlrd==0.7.9
xlwt==0.7.4
django-tinymce==2.3.0
django-htmlmin==0.9.1
awesome-slugify==1.6.5